Add tests for checking RDFlib consistency
Keep canonized editions from using different RDFlib versions on the same
input file and compare the content in a test.
This ensures that if there are any inconsistencies between the RDFlib
versions used, they should be picked up on - note, the weak link here is
the base input file. If it does not contain enough edge cases, this is
not trustworthy.