Test more exceptions - update canonize workflow
Update the canonize workflow by adding in explicit triple sorting.
The triples from the parsed ontology is sorted and re-added to a new
`rdflib.Graph`, reusing the namespaces and base.
Then after this, the ontology is exported.

Test and expand on validation exceptions and warnings.
Check readability as well as writeability.