Generate ontology from different sources and test
An ontology generated using EMMOntoPy has been added, opened in Protégé
and saved as a Turtle file.
These files are put in a special folder where future other generation
sources that generate a Turtle file from the same ontology can be put.
All files in this folder are canonized in a test.

The ultimate source of truth is the set of triples in each file (for
canonization).
At a later stage a `diff`-like tool can be implemented that can relate
the difference in triples to each other between samey ontologies.