New CLI option: --fail-fast
Wrap canonizing several files in a while loop, to be able to cache
errors and warnings to print in a summary at the end.
Unless `--fail-fast` is specified, in which case it will print a single
ERROR as soon as an error occurs.

If only a single file is given and it provides a warning a standard
WARNING is printed, instead of the new summary.